refactor!: monorepo (apps/web) consuming @kotahusky/ui
Move the Next app into apps/web; root becomes the npm-workspaces + turbo
orchestrator. apps/web imports components from @kotahusky/ui and supplies
app-specific props (carousel images, telegram copy, blue aurora palette);
global.css imports the package theme, @source-scans it, and pins brand->blue.
Local component copies removed. Dockerfile rewritten for the Next standalone
monorepo layout (CMD node apps/web/server.js); release versions apps/web.

Verified: turbo build + test + lint green; brand-* utilities and text-shadow-soft
generated in the production CSS. NOT verified: docker image build (no daemon local).

BREAKING CHANGE: repo layout is now a monorepo; app lives in apps/web.
