--- app/lib/media.ts
@@ -1,23 +1,0 @@
-import { useSyncExternalStore } from 'react';
-
-const emptySubscribe = () => () => {};
-
-/** True only after client hydration — avoids SSR/CSR markup mismatches. */
-export function useIsClient(): boolean {
-  return useSyncExternalStore(emptySubscribe, () => true, () => false);
-}
-
-function subscribeReducedMotion(callback: () => void) {
-  const mq = window.matchMedia('(prefers-reduced-motion: reduce)');
-  mq.addEventListener('change', callback);
-  return () => mq.removeEventListener('change', callback);
-}
-
-/** Tracks the user's `prefers-reduced-motion` setting (false during SSR). */
-export function usePrefersReducedMotion(): boolean {
-  return useSyncExternalStore(
-    subscribeReducedMotion,
-    () => window.matchMedia('(prefers-reduced-motion: reduce)').matches,
-    () => false,
-  );
-}
