+++ apps/web/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "web",
+  "version": "2.2.2",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "serve": "next start",
+    "lint": "eslint .",
+    "test": "vitest run",
+    "test:watch": "vitest"
+  },
+  "dependencies": {
+    "@heroicons/react": "^2.2.0",
+    "@kotahusky/ui": "*",
+    "@nauverse/react-aurora-background": "^1.0.12",
+    "next": "^16.2.9",
+    "qrcode.react": "^4.2.0",
+    "react": "^19.2.7",
+    "react-dom": "^19.2.7",
+    "react-icons": "^5.6.0",
+    "tslib": "^2.8.1"
+  }
+}
