+++ apps/web/staticwebapp.config.json
