--- next.config.js
@@ -1,10 +1,0 @@
-//@ts-check
-
-/**
- * @type {import('next').NextConfig}
- **/
-const nextConfig = {
-  output: 'standalone',
-};
-
-module.exports = nextConfig;
