+++ package.json
@@ -1,39 +1,27 @@
 {
-  "name": "homepage",
-  "version": "2.2.2",
+  "name": "homepage-monorepo",
+  "version": "0.0.0",
   "license": "MIT",
-  "scripts": {
-    "dev": "next dev",
-    "serve": "npx next start",
-    "build": "next build",
-    "lint": "eslint .",
-    "start": "npx next start",
-    "start:docker-clean-build": "docker build -t homepage:latest .;docker run -p 3000:3000 homepage:latest",
-    "test": "vitest run",
-    "test:watch": "vitest"
-  },
-  "packageManager": "npm@9.6.7",
   "private": true,
+  "packageManager": "npm@9.6.7",
   "workspaces": [
+    "apps/*",
     "packages/*"
   ],
+  "scripts": {
+    "dev": "npm run dev -w web",
+    "build": "turbo run build",
+    "start": "npm run start -w web",
+    "lint": "turbo run lint",
+    "test": "turbo run test"
+  },
   "engines": {
     "node": ">=24.0.0"
   },
-  "dependencies": {
-    "@heroicons/react": "^2.2.0",
-    "@nauverse/react-aurora-background": "^1.0.12",
-    "next": "^16.2.9",
-    "qrcode.react": "^4.2.0",
-    "react": "^19.2.7",
-    "react-dom": "^19.2.7",
-    "react-icons": "^5.6.0",
-    "tslib": "^2.8.1"
-  },
   "overrides": {
     "@nauverse/react-aurora-background": {
-      "react": "$react",
-      "react-dom": "$react-dom"
+      "react": "^19.2.7",
+      "react-dom": "^19.2.7"
     }
   },
   "devDependencies": {
